Caja y cierre ticket OK

+++ src/main/java/es/udc/fi/tfg/seguimiento/model/LineaTicket.java
@@ -20,7 +20,7 @@ public class LineaTicket {
 
 	private Long idLineaTicket;
 	private Integer cantidad;
-	private Float precio;
+	private Double precio;
 	private Integer iva;
 	private Producto producto;
 	private Ticket ticket;
@@ -29,7 +29,7 @@ public class LineaTicket {
 		
 	}
 	
-	public LineaTicket(Integer cantidad, Float precio, Integer iva, Producto producto, Ticket ticket) {
+	public LineaTicket(Integer cantidad, Double precio, Integer iva, Producto producto, Ticket ticket) {
 		super();
 		this.cantidad = cantidad;
 		this.precio = precio;
@@ -60,11 +60,11 @@ public class LineaTicket {
 	}
 
 	@Column (name="precio")
-	public Float getPrecio() {
+	public Double getPrecio() {
 		return precio;
 	}
 
-	public void setPrecio(Float precio) {
+	public void setPrecio(Double precio) {
 		this.precio = precio;
 	}
